fixed a bug in the initialisation of gridworld

+++ src/models/PolicyRewardBelief.cc
@@ -127,8 +127,8 @@ FixedDiscretePolicy* PolicyRewardBelief::getPolicy()
     int n_samples = betas.size();
     real sum = 0.0;
     Matrix R(n_states, n_actions);
-    assert(n_samples == rewards.size());
-    assert(n_samples == sample_counts.size());
+    assert(n_samples == (int) rewards.size());
+    assert(n_samples == (int) sample_counts.size());
     for (int i=0; i<n_samples; ++i) {
         R += rewards[i] * sample_counts[i];
         sum += sample_counts[i];
